Added texture loading with coroutines, didn't test

+++ WinD3D12/Buffers.h
@@ -39,7 +39,7 @@ namespace ver
 			return count;
 		}
 	private:
-		uint32_t count;
+		size_t count;
 		wgpu::IndexFormat format;
 	};
 	class VertexBuffer : public Buffer
